@@ -1,4 +1,5 @@
 {
+	"type": "module",
 	"name": "inkshio-start-project",
 	"version": "2.0.0",
 	"description": "Сборка проекта",
@@ -45,25 +46,25 @@
 		"url": "https://github.com/inkShio/inkShio-start-project.git"
 	},
 	"devDependencies": {
-		"@babel/core": "^7.19.3",
-		"@babel/preset-env": "^7.19.4",
-		"@babel/register": "^7.18.9",
+		"@babel/core": "^7.23.6",
+		"@babel/preset-env": "^7.23.6",
+		"@babel/register": "^7.22.15",
 		"@emitty/core": "^2.0.1",
 		"@emitty/language-pug": "^2.0.1",
 		"babelify": "^10.0.0",
-		"browser-sync": "^2.27.10",
+		"browser-sync": "^2.29.3",
 		"browserify": "^17.0.0",
-		"del": "^6.1.1",
-		"eslint": "^8.25.0",
+		"del": "^7.1.0",
+		"eslint": "^8.56.0",
 		"eslint-config-airbnb-base": "^15.0.0",
-		"eslint-plugin-import": "^2.26.0",
+		"eslint-plugin-import": "^2.29.1",
 		"gulp": "^4.0.2",
-		"gulp-autoprefixer": "^8.0.0",
+		"gulp-autoprefixer": "^9.0.0",
 		"gulp-cheerio": "^1.0.0",
 		"gulp-clean-css": "^4.3.0",
 		"gulp-concat": "^2.6.1",
 		"gulp-data": "^1.3.1",
-		"gulp-debug": "^4.0.0",
+		"gulp-debug": "^5.0.1",
 		"gulp-fonter": "^0.3.0",
 		"gulp-group-css-media-queries": "^1.2.2",
 		"gulp-if": "^3.0.0",
@@ -77,7 +78,7 @@
 		"gulp-pug": "^5.0.0",
 		"gulp-real-favicon": "^0.3.2",
 		"gulp-rename": "^2.0.0",
-		"gulp-replace": "^1.1.3",
+		"gulp-replace": "^1.1.4",
 		"gulp-sass": "^5.1.0",
 		"gulp-sass-glob": "^1.1.0",
 		"gulp-sourcemaps": "^3.0.0",
@@ -87,28 +88,28 @@
 		"gulp-uglify": "^3.0.2",
 		"gulp-version-number": "^0.2.4",
 		"gulp-webpcss": "^1.1.1",
-		"husky": "^8.0.1",
+		"husky": "^8.0.3",
 		"imagemin-pngquant": "^9.0.2",
 		"imagemin-webp": "^6.0.0",
-		"lint-staged": "^13.0.3",
+		"lint-staged": "^15.2.0",
 		"posthtml-attrs-sorter": "^1.1.0",
-		"pug-lint": "^2.6.0",
-		"sass": "^1.55.0",
-		"stylelint": "^14.14.0",
+		"pug-lint": "^2.7.0",
+		"sass": "^1.69.5",
+		"stylelint": "^16.1.0",
 		"stylelint-config-rational-order": "^0.1.2",
-		"stylelint-config-standard": "^29.0.0",
-		"stylelint-order": "^5.0.0",
-		"stylelint-scss": "^4.3.0",
+		"stylelint-config-standard": "^36.0.0",
+		"stylelint-order": "^6.0.4",
+		"stylelint-scss": "^6.0.0",
 		"through2": "^4.0.2",
 		"vinyl-buffer": "^1.0.1",
 		"vinyl-source-stream": "^2.0.0",
 		"webp-converter": "^2.2.3"
 	},
 	"dependencies": {
-		"bootstrap": "^5.2.2",
+		"bootstrap": "^5.3.2",
 		"focus-visible": "^5.2.0",
 		"ink-reset-scss": "^2.0.0",
-		"jquery": "^3.6.1",
+		"jquery": "^3.7.1",
 		"object-fit-images": "^3.2.4",
 		"sass-mediaqueries": "^1.6.1"
 	}
