@@ -1,12 +1,13 @@
-import del from 'del';
-import fs from "fs";
+// import del from 'del';
+import { deleteAsync } from 'del';
+import fs from 'fs';
 import gulp from 'gulp';
 import debug from 'gulp-debug';
 import fonter from 'gulp-fonter';
 import plumber from 'gulp-plumber';
 import ttf2woff2 from 'gulp-ttf2woff2';
-import config from '../config';
-import { scssBuild } from './styles';
+import config from '../config.js';
+import { scssBuild } from './styles.js';
 
 export const otfToTtf = () => (
 	gulp.src(`${config.paths.fonts.app}/*.otf`)
@@ -99,8 +100,6 @@ export const fontsStyle = (cb) => {
 	cb();
 }
 
-const cleanFonts = () => (
-	del([config.paths.fonts.build, `${config.paths.root.tmp}/*.ttf`, `${config.paths.root.tmp}/fonts-generated.scss`])
-);
+const cleanFonts = await deleteAsync([config.paths.fonts.build, `${config.paths.root.tmp}/*.ttf`, `${config.paths.root.tmp}/fonts-generated.scss`]);
 
 export const fontsWatch = () => gulp.watch(config.paths.fonts.watch, gulp.series(cleanFonts, otfToTtf, ttfToWoff, ttfToWoff2, fontsStyle, scssBuild));
